test: split quote auth csrf transport tests
